requirements.txt
Django==5.2.5
Pillow==10.4.0
PyMySQL==1.1.1
mysqlclient==2.2.4
djangorestframework==3.16.1
requests==2.31.0
requests-oauthlib==1.3.1
cryptography==43.0.1
django-cors-headers==4.3.1
python-decouple==3.8
